# docs/baseline/openapi.yaml
openapi: 3.1.0
info:
  title: Gamja Market API
  version: 0.9.0
  description: Account, product, chat, safety, trade, configurable Toss payment, durable webhook, settlement, and administrator operations contract. Authentication uses an opaque HttpOnly session cookie.
servers:
  - url: /api/v1
paths:
...
  /auth/signups:
    post:
      operationId: signUp
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SignUpRequest' } } } }
      responses:
        '201': { description: Contact confirmation was issued. }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/contact-confirmations:
    post:
      operationId: confirmContact
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CodeConfirmationRequest' } } } }
      responses:
        '201': { description: Account is active. }
        '400': { $ref: '#/components/responses/Problem' }
  /auth/contact-confirmations/request:
    post:
      operationId: resendContactConfirmation
      description: Reissues a contact verification through the configured delivery provider without revealing account existence.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/IdentifierRequest' } } } }
      responses:
        '202': { description: Always accepted when not rate limited. }
        '429': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /auth/login:
    post:
      operationId: login
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }
      responses:
        '201':
          description: Session is issued in the gm_session HttpOnly cookie.
          headers:
            Set-Cookie: { schema: { type: string } }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      operationId: logout
      security: [{ sessionCookie: [] }]
      responses:
        '204': { description: Session revoked. }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/me:
    get:
      operationId: getCurrentAccount
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Current account. }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/password-resets:
    post:
      operationId: requestPasswordReset
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/IdentifierRequest' } } } }
      responses:
        '202': { description: Always accepted to avoid account enumeration. }
  /auth/password-resets/confirm:
    post:
      operationId: confirmPasswordReset
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetConfirmRequest' } } } }
      responses:
        '204': { description: Password changed and sessions revoked. }
        '400': { $ref: '#/components/responses/Problem' }
  /auth/me/neighborhood:
    patch:
      operationId: setCurrentNeighborhood
      security: [{ sessionCookie: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/NeighborhoodSelection' } } } }
      responses:
        '200': { description: Selected administrative neighborhood. }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/me/withdrawal:
    post:
      operationId: withdrawCurrentAccount
      security: [{ sessionCookie: [] }]
      responses:
        '201': { description: Account is withdrawn and sessions are revoked. }
        '401': { $ref: '#/components/responses/Problem' }
...
  /reports:
    post:
      operationId: createReport
      security: [{ sessionCookie: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateReportRequest' } } } }
      responses:
        '201': { description: Report received without exposing internal investigation data. }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
  /trades:
    get:
      operationId: listTrades
      description: Lists trades where the current account is the buyer or seller.
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Participant-visible trades with KRW integer strings and the current account role. }
        '401': { $ref: '#/components/responses/Problem' }
    post:
      operationId: requestTrade
      description: Requests a trade and creates or returns its buyer-seller product chat in the same transaction after product, neighborhood, ownership, duplicate, and block checks.
      security: [{ sessionCookie: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateTradeRequest' } } } }
      responses:
        '201': { description: Trade requested with a server-side price snapshot and chatId. }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      description: SUPER_ADMIN-only aggregate work queue counts.
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: User, product, open report, and active trade counts. }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /admin/users:
    get:
      operationId: listAdminUsers
      description: SUPER_ADMIN-only masked user list.
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Up to 100 users with masked email and phone values. }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /admin/reports:
    get:
      operationId: listAdminReports
      description: SUPER_ADMIN-only report queue.
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Up to 100 reports ordered by creation time. }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /admin/trades:
    get:
      operationId: listAdminTrades
      description: SUPER_ADMIN-only trade operations view.
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Up to 100 trades with product summaries and KRW integer strings. }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /admin/audit-logs:
    get:
      operationId: listAdminAuditLogs
      description: SUPER_ADMIN-only immutable audit log view.
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Up to 200 audit records ordered by occurrence time. }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /admin/payments:
    get:
      operationId: listAdminPayments
      description: SUPER_ADMIN-only payment, settlement, and recent webhook operations view with masked provider identifiers.
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Up to 100 recent payments. }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /admin/outbox-events:
    get:
      operationId: listFailedOutboxEvents
      description: SUPER_ADMIN-only failed and dead-letter durable event queue.
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Up to 100 failed durable events. }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /admin/outbox-events/{eventId}/retry:
    post:
      operationId: retryOutboxEvent
      description: SUPER_ADMIN-only audited retry of a failed or dead-letter durable event.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: eventId, in: path, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminActionRequest' } } } }
      responses:
        '201': { description: Event queued for a fresh audited retry. }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /admin/payments/{paymentId}/reconcile:
    post:
      operationId: reconcileAdminPayment
      description: SUPER_ADMIN-only provider lookup and audited comparison with internal payment state.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: paymentId, in: path, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminActionRequest' } } } }
      responses:
        '201': { description: Provider lookup comparison and audit record completed. }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /admin/users/{userId}/suspend:
    post:
      operationId: suspendAdminUser
      description: Suspends a non-admin user after SUPER_ADMIN password reauthentication and records an audit reason.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminActionRequest' } } } }
      responses:
        '201': { description: User suspended and active sessions revoked. }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /admin/users/{userId}/activate:
    post:
      operationId: activateAdminUser
      description: Reactivates a non-admin user after SUPER_ADMIN password reauthentication and records an audit reason.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminActionRequest' } } } }
      responses:
        '201': { description: User activated. }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /admin/products/{productId}/hide:
    post:
      operationId: hideAdminProduct
      description: Hides a product after SUPER_ADMIN password reauthentication and records before and after values.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: productId, in: path, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminActionRequest' } } } }
      responses:
        '201': { description: Product hidden. }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /admin/reports/{reportId}/resolve:
    post:
      operationId: resolveAdminReport
      description: Resolves a report after SUPER_ADMIN password reauthentication and records an audit reason.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: reportId, in: path, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminActionRequest' } } } }
      responses:
        '201': { description: Report resolved. }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /admin/reports/{reportId}/dismiss:
    post:
      operationId: dismissAdminReport
      description: Dismisses a report after SUPER_ADMIN password reauthentication and records an audit reason.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: reportId, in: path, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminActionRequest' } } } }
      responses:
        '201': { description: Report dismissed. }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: gm_session }
  schemas:
...
    SignUpRequest:
      type: object
      required: [password, adultConfirmed]
      properties:
        email: { type: string, format: email }
        phone: { type: string }
        password:
          type: string
          minLength: 12
          maxLength: 128
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[^A-Za-z0-9\s]).{12,128}$'
          description: Server additionally rejects common or predictable strings, four repeated characters, and account identifier inclusion.
          writeOnly: true
        adultConfirmed: { type: boolean, const: true }
    IdentifierRequest:
      type: object
      required: [identifier]
      properties: { identifier: { type: string } }
    CodeConfirmationRequest:
      allOf:
        - $ref: '#/components/schemas/IdentifierRequest'
        - type: object
          required: [code]
          properties: { code: { type: string, minLength: 6, maxLength: 6, writeOnly: true } }
    PasswordResetConfirmRequest:
      allOf:
        - $ref: '#/components/schemas/CodeConfirmationRequest'
        - type: object
          required: [newPassword]
          properties:
            newPassword:
              type: string
              minLength: 12
              maxLength: 128
              pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[^A-Za-z0-9\s]).{12,128}$'
              description: Uses the same server-enforced strength policy as signup.
              writeOnly: true
    LoginRequest:
      allOf:
        - $ref: '#/components/schemas/IdentifierRequest'
        - type: object
          required: [password]
          properties: { password: { type: string, writeOnly: true } }
    NeighborhoodSelection:
      type: object
      required: [neighborhoodCode]
      properties: { neighborhoodCode: { type: string } }
...
    CreateReportRequest:
      type: object
      additionalProperties: false
      required: [targetType, targetId, reason]
      properties:
        targetType: { type: string, enum: [USER, PRODUCT, MESSAGE] }
        targetId: { type: string, minLength: 1 }
        reason: { type: string, minLength: 2, maxLength: 100 }
        detail: { type: string, maxLength: 2000 }
    CreateTradeRequest:
      type: object
      additionalProperties: false
      required: [productId]
      properties:
        productId: { type: string, minLength: 1 }
...
    AdminActionRequest:
      type: object
      additionalProperties: false
      required: [reason, password]
      properties:
        reason: { type: string, minLength: 2, maxLength: 500 }
        password: { type: string, minLength: 12, maxLength: 200, writeOnly: true }
...
    Problem:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        requestId: { type: string }
  responses:
    Problem:
      description: Safe problem response.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }